// api/client/models/OneTimePasscodeVerifyRequest.ts
/**
 * api.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* request
*/
export class OneTimePasscodeVerifyRequest {
    'emailOrPhone'?: string;
    'code'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "emailOrPhone",
            "baseName": "emailOrPhone",
            "type": "string",
            "format": ""
        },
        {
            "name": "code",
            "baseName": "code",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return OneTimePasscodeVerifyRequest.attributeTypeMap;
    }
    
    public constructor() {
    }
}