// api/client/models/PlaidCreateLinkTokenResponse.ts
/**
 * Flux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: evergreen
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* response
*/
export class PlaidCreateLinkTokenResponse {
    'linkToken'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "linkToken",
            "baseName": "linkToken",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return PlaidCreateLinkTokenResponse.attributeTypeMap;
    }
    
    public constructor() {
    }
}